+++ botfront/client/main.less
@@ -6,15 +6,13 @@
 @import '../imports/ui/components/nlu/models/style/NLUModel.import.less';
 @import '../imports/ui/components/nlu/models/style/style.import.less';
 @import '../imports/ui/components/templates/template-upsert/NLUCriteria/NLUCriteriaEditor.import.less';
-@import '../imports/ui/components/utils/style/ImageField.import.less';
-@import '../imports/ui/components/utils/style/style.import.less';
+@import '../imports/ui/components/utils/style.import.less';
 @import '../imports/ui/components/nlu/activity/style.import.less';
 @import '../imports/ui/components/templates/template-upsert/Template.import.less';
 @import '../imports/ui/components/templates/templates-list/style.import.less';
 @import '../imports/ui/components/common/style.import.less';
-@import '../imports/ui/components/nlu/common/EntityDropdown.import.less';
 @import '../imports/ui/components/stories/StoriesFooter.import.less';
-@import '../imports/ui/components/nlu/common/IntentDropdown.import.less';
+@import '../imports/ui/components/nlu/common/style.import.less';
 @import '../imports/ui/components/conversations/style.import.less';
 @import '../imports/ui/components/stories/style.import.less';
 @import '../imports/ui/components/account/style.import.less';
