--- botfront/imports/ui/components/nlu/common/EntityDropdown.import.less
@@ -1,17 +1,0 @@
-.entity-dropdown.ui.selection.dropdown {
-  border-color: fade(@teal, 50%);
-  border-top: none;
-  border-right: none;
-  border-left: none;
-  border-radius: 0;
-  min-width: 12em;
-}
-
-.entity-dropdown.ui.selection.dropdown>input.search {
-  padding-left: 3em !important;
-}
-
-.entity-dropdown.ui.selection.dropdown.icon>.icon {
-  background-color: fade(@teal, 5%);
-  color: @teal;
-}
