--- botfront/imports/ui/components/nlu/common/IntentDropdown.import.less
@@ -1,17 +1,0 @@
-.intent-dropdown.ui.selection.dropdown {
-  border-color: fade(@purple, 50%);
-  border-top: none;
-  border-right: none;
-  border-left: none;
-  border-radius: 0;
-  min-width: 12em;
-}
-
-.intent-dropdown.ui.selection.dropdown > input.search {
-  padding-left: 3em !important;
-}
-
-.intent-dropdown.ui.selection.dropdown.icon>.icon {
-  background-color: fade(@purple, 5%);
-  color: @purple;
-}
