+++ botfront/imports/ui/components/settings/Instances.jsx
@@ -6,9 +6,9 @@ import {
     AutoField, AutoForm, HiddenField, SubmitField,
 } from 'uniforms-semantic';
 import Alert from 'react-s-alert';
-import { InstanceSchema } from '../../../../api/instances/instances.schema';
-import { Instances as InstancesCollection } from '../../../../api/instances/instances.collection';
-import { wrapMeteorCallback } from '../../utils/Errors';
+import { InstanceSchema } from '../../../api/instances/instances.schema';
+import { Instances as InstancesCollection } from '../../../api/instances/instances.collection';
+import { wrapMeteorCallback } from '../utils/Errors';
 
 class Instances extends React.Component {
     onValidate = (model, error, callback) => {
