+++ botfront/imports/ui/components/stories/StoryGroupBrowser.jsx
@@ -7,7 +7,7 @@ import { withTracker } from 'meteor/react-meteor-data';
 import { connect } from 'react-redux';
 import { Slots } from '../../../api/slots/slots.collection';
 import StoryGroupItem from './StoryGroupItem';
-import { ConversationOptionsContext } from '../utils/Context';
+import { ConversationOptionsContext } from './Context';
 
 class StoryGroupBrowser extends React.Component {
     constructor(props) {
