+++ botfront/imports/ui/components/stories/common/UserUtteranceContainer.jsx
@@ -4,8 +4,8 @@ import React, {
 import PropTypes from 'prop-types';
 import { Button } from 'semantic-ui-react';
 import { OOS_LABEL } from '../../constants.json';
-import FloatingIconButton from '../../nlu/common/FloatingIconButton';
-import UserUtteranceViewer from '../../utils/UserUtteranceViewer';
+import FloatingIconButton from '../../common/FloatingIconButton';
+import UserUtteranceViewer from '../../nlu/common/UserUtteranceViewer';
 import { ProjectContext } from '../../../layouts/context';
 import UtteranceInput from '../../utils/UtteranceInput';
 import ExceptionWrapper from './ExceptionWrapper';
