+++ botfront/imports/ui/components/utils/style.import.less
@@ -1,0 +1,47 @@
+div.image-field {
+  width: fit-content;
+  padding: 10px;
+
+  .cancel-button {
+    float: right;
+  }
+
+  img {
+    max-height: 200px;
+  }
+}
+
+.trigger {
+  display: inline;
+
+  .input-container {
+    display: inline;
+  }
+}
+
+.inline_search {
+  .item.search-menu-item {
+    white-space: nowrap;
+    font-size: 12px;
+    height: 26px;
+    padding: 7px;
+  }
+
+  .vertical.menu {
+    width: 13rem;
+    min-height: 0;
+    max-height: 132px;
+    overflow-y: scroll;
+  }
+}
+
+.hidden-component {
+  display: block;
+}
+
+.ui.form div.field {
+  label {
+    display: inline-block;
+    margin-right: 4px;
+  }
+}
