--- botfront/imports/ui/components/utils/style/ImageField.import.less
@@ -1,12 +1,0 @@
-div.image-field {
-  width: fit-content;
-  padding: 10px;
-
-  .cancel-button {
-    float: right;
-  }
-
-  img {
-    max-height: 200px;
-  }
-}
