--- botfront/imports/ui/components/utils/style/style.import.less
@@ -1,92 +1,0 @@
-.trigger {
-  display: inline;
-
-  .input-container {
-    display: inline;
-  }
-}
-
-.inline_search {
-  .item.search-menu-item {
-    white-space: nowrap;
-    font-size: 12px;
-    height: 26px;
-    padding: 7px;
-  }
-
-  .vertical.menu {
-    width: 13rem;
-    min-height: 0;
-    max-height: 132px;
-    overflow-y: scroll;
-  }
-}
-
-.hidden-component {
-  display: block;
-}
-
-.ui.form div.field {
-  label {
-    display: inline-block;
-    margin-right: 4px;
-  }
-}
-
-.entity-container {
-  line-height: 1.25;
-  display: inline-block;
-  position: relative;
-  .float {
-    position: absolute;
-    width: 100px;
-    bottom: ~"calc(100% - 1px)";
-    font-size: small;
-  }
-}
-
-.createentitycolorclasses(@i:1) when(@i <= length(@basic-colors)) {
-  @name: extract(extract(@basic-colors, @i), 1);
-  @color: extract(extract(@basic-colors, @i), 2);
-
-  .entity-container.@{name} {
-    :not(.float) {
-      background-color: fade(@color, 17%);
-      border-bottom: 2.5px solid @color;
-    }
-    .float {
-      color: @color
-    }
-  }
-  .createentitycolorclasses(@i + 1);
-}
-
-.createentitycolorclasses();
-
-.utterance-viewer {
-  .selected-text {
-    background-color: #c4d8f5;
-  }
-}
-
-.intent-label {
-  display: inline-block;
-  color: @purple;
-  width: max-content;
-  font-weight: medium;
-  padding: 2px;
-  margin: 0;
-  cursor: default;
-
-  &.disabled {
-    color: @grey;
-  }
-  &.null {
-    color: darken(@purple, 20%);
-  }
-}
-
-.intent-dropdown-container {
-  display: flex;
-  width: max-content;
-}
